Improve construction of viewing conditions
and be more consitent with floats, illuminants and stuff.

+++ de.fhg.igd.pcolor/src/de/fhg/igd/pcolor/colorspace/CS_JCh.java
@@ -58,7 +58,7 @@ public class CS_JCh extends CS_CIECAM02 {
 	 * @param Y_b adaptation luminance of color background
 	 * @param sur surrounding
 	 */
-	public CS_JCh(double[] XYZWhitePoint, double L_A, double Y_b, Surrounding sur) {
+	public CS_JCh(float[] XYZWhitePoint, double L_A, double Y_b, Surrounding sur) {
 		super(XYZWhitePoint, L_A, Y_b, sur);
 	}
